Add a few other classes to make the code more oop
